ResMgr: Change request and ack structures
Now, all the resources are requested by name. The translation from the
resource id to the name is done in the IpcResource.c module. That
means that the exported APIs remain the same.

Also, request and ack structure are split into sub structures depending
on the action requested or the ack of a specific action. That is because
each action needs its own parameters.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>