ResMgr: Add local structures for auxclk and regulator resources
Now the auxclk resource receives the name of which auxclk is requested
as a  parameter (to avoid limit the # of auxclk that can be request
and to avoid hardcoding that in the Host side). Therefore, now the
auxclk structure is different and we need to create a local auxclk
structure match the Host side and keep the original structure to match
exported APIs. The same applies to the regulator resource.

The old structures can be removed once all users of IpcResource are
migrated to the new changes.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>