IPC: VirtQueue: simplify init sequence
We don't really need the physical address of the vrings, since we
access them through the IOMMU anyway, and therefore only interested
in device addresses (those that goes to the IOMMU).

It's perfectly fine to hard code device address, since it's our address
space, and this is actually what we've been doing.

This patch:
1. renames the vring address macros (to prevent phys/device address
   confusion)
2. remove the initStage and the semaphone init sequence code

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>