IPC: Correct VirtQueue_getAvailBuf based on review of virtio lguest
A review of VirtQueue revealed the index to the available ring buffer was
being calculated differently than lguest from which it was modelled.
It worked given the current usage model, but to be more correct, it
should be updated to match what Linux side does.

lguest.c -> wait_for_vq_desc:

    head = vq->vring.avail->ring[last_avail % vq->vring.num];
    lg_last_avail(vq)++;

VirtQueue_getAvailBuf (previously):
    head = vq->last_avail_idx++ % vq->vring.num;

This change was verified to work by asserting that head index was the same
in both cases under sample tests.

Signed-off-by: G Anthony <a0783926@ti.com>